+++ Source/DataTransfer.Legacy.MassImport/RelEyeTelemetry/Events/EventJobStart.cs
@@ -1,0 +1,13 @@
+﻿// <copyright file="EventJobStart.cs" company="Relativity ODA LLC">
+// © Relativity All Rights Reserved.
+// </copyright>
+
+namespace DataTransfer.Legacy.MassImport.RelEyeTelemetry.Events
+{
+	/// <inheritdoc/>
+	public class EventJobStart : EventBase
+	{
+		/// <inheritdoc/>
+		public override string EventName => Constants.EventName.JobStart;
+	}
+}